workaround: ipython tab-complete
the `ctrl` object does not have ipython tab complete functionality any
more; it works as long as it does not inherit from `CHIRPBroadcaster`
class. There seem to be several issues at play and a fix does not seem
straightforward. Instead, the `ctrl.constellation` object is exposed at
the iPython console as `constellation` which provides tab complete with
less typing!