python/core/satellite: remove timeout on stop
Would previously kill the `do_run` thread; however, there is no
technical need to force the user to perform stopping actions only in
`do_stopping`. Removing the timeout allows to either wrap things in in
`do_run` alone, only in `do_stopping` or both. The FSM will be in
`STOPPING` the entire time, so the process is transparent to the system.