cxx/logging: only enable CMDP after name is set
By default, the global CMDP log level is OFF, meaning no messages are forwarded to the CMDP sink. Only once the subscription thread starts, this log level will be adjusted. Thus the subscription thread is now started in the `enableSending` method of the sink, which also sets the name and registers the CHIRP service.

Signed-off-by: Stephan Lachnit <stephanlachnit@debian.org>